% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gproj.R
\name{detect_rectilinear}
\alias{detect_rectilinear}
\title{Test if 2D coordinate arrays are rectilinear (lon varies only along cols,
lat varies only along rows)}
\usage{
detect_rectilinear(lon, lat, tol = 0.01)
}
\arguments{
\item{lon}{2D matrix of longitudes (nrow x ncol)}

\item{lat}{2D matrix of latitudes (nrow x ncol)}

\item{tol}{tolerance for "constant along an axis" test (in degrees)}
}
\value{
list describing the rectilinear structure
}
\description{
Test if 2D coordinate arrays are rectilinear (lon varies only along cols,
lat varies only along rows)
}
