% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gproj-package.R
\docType{package}
\name{gproj-package}
\alias{gproj}
\alias{gproj-package}
\title{gproj: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/gproj}
  \item Report bugs at \url{https://github.com/mdsumner/gproj/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
