% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gproj.R
\name{grid_spec}
\alias{grid_spec}
\title{Convenience: generate a grid specification from a gproj_region result}
\usage{
grid_spec(region, resolution_m = 1000)
}
\arguments{
\item{region}{result from gproj_region()}

\item{resolution_m}{grid cell size in metres}
}
\value{
list with crs, extent, dimension, resolution
}
\description{
Convenience: generate a grid specification from a gproj_region result
}
