% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gproj.R
\name{initial_bearing}
\alias{initial_bearing}
\title{Initial bearing (forward azimuth) from point 1 to point 2
All inputs in degrees, output in degrees [0, 360)}
\usage{
initial_bearing(lon1, lat1, lon2, lat2)
}
\description{
Initial bearing (forward azimuth) from point 1 to point 2
All inputs in degrees, output in degrees [0, 360)
}
