% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gproj.R
\name{proj}
\alias{proj}
\title{Build a PROJ string from params and a projection family or property}
\usage{
proj(p, family = "auto")
}
\arguments{
\item{p}{a params object}

\item{family}{projection family name (e.g. "laea", "stere", "omerc", "lcc", "ortho")
or a property: "equal_area", "equidistant", "conformal", "auto"}
}
\description{
Build a PROJ string from params and a projection family or property
}
