% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gproj.R
\name{test_regular_spacing}
\alias{test_regular_spacing}
\title{Test whether a 1D coordinate sequence is uniformly spaced (regular)}
\usage{
test_regular_spacing(values, tol_relative = 0.001)
}
\arguments{
\item{values}{numeric vector}

\item{tol_relative}{relative tolerance}
}
\value{
list with is_regular, resolution, max_deviation
}
\description{
Test whether a 1D coordinate sequence is uniformly spaced (regular)
}
