% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gproj-package.R
\name{viirs_eg}
\alias{viirs_eg}
\title{viirs_swaths'
An example swath source}
\usage{
viirs_eg()
}
\value{
string to a url with vsicurl protocol
}
\description{
Only for use in examples in this package.
}
\examples{
viirs_eg() ## open this with 'new(gdalraster::GDALRaster, <>)' or terra::rast
}
