% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gproj-package.R
\docType{package}
\name{gproj-package}
\alias{gproj}
\alias{gproj-package}
\title{gproj: Projection Concierge}
\description{
A 'Projection Concierge': given coordinates or a region of interest, it tells you what projection to use and gives you the transform.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/gproj}
  \item Report bugs at \url{https://github.com/mdsumner/gproj/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
